# perf: improve alignment::distance::levenshtein and alignment::distance::bounded_levenshtein on strrings where distance is small