--- a/.github/workflows/rust.yml
+++ b/.github/workflows/rust.yml
@@ -86,7 +86,7 @@ jobs:
     needs: Formatting
     runs-on: ubuntu-latest
     env:
-      MSRV_VERSION: 1.60.0
+      MSRV_VERSION: 1.62.0
     steps:
       - name: Checkout repository
         uses: actions/checkout@v2