--- a/Cargo.toml
+++ b/Cargo.toml
@@ -53,6 +53,7 @@ triple_accel = ">=0.3, <0.5"
 thiserror = "1"
 anyhow = "1"
 rand = ">=0.7.3, < 0.9"
+editdistancek = "1"
 
 [dependencies.vec_map]
 version = "0.8"