--- a/README.md
+++ b/README.md
@@ -57,7 +57,7 @@ For extra credit, feel free to familiarize yourself with:
 
 ## Minimum supported Rust version
 
-Currently the minimum supported Rust version is 1.60.0.
+Currently the minimum supported Rust version is 1.62.0.
 
 ## License