--- a/src/alignment/distance.rs
+++ b/src/alignment/distance.rs
@@ -164,7 +164,6 @@ pub mod simd {
     /// assert_eq!(ldist, None);
     /// ```
     pub fn bounded_levenshtein(alpha: TextSlice<'_>, beta: TextSlice<'_>, k: u32) -> Option<u32> {
-        // triple_accel::levenshtein::levenshtein_simd_k(alpha, beta, k)
         if let Some(x) = editdistancek::edit_distance_bounded(alpha, beta, min(k as usize, max(alpha.len(), beta.len()))) {
             Some(x as u32)
         } else {