@RagnarGrootKoerkamp 2023-06-29
This doesn't compile on macbook M1.
https://github.com/RagnarGrootKoerkamp/astar-pairwise-aligner/issues/17